search profil

+++ database/migrations/2017_02_24_125945_create_profil_table.php
@@ -22,7 +22,7 @@ class CreateProfilTable extends Migration
             $table->integer('cheveux');
             $table->integer('yeux');
             $table->integer('poid');
-            $table->integer('taille');
+            $table->float('taille',3,2);
             $table->integer('users_id');
             $table->string('pseudo');
             $table->string('mail');
